pyproject.toml
[project]
name = "intelligence"
version = "0.1.0"
description = ""
authors = [
    {name = "Husani Oakley",email = "husani.oakley@husani.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "authlib (>=1.6.9,<2.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "anthropic (>=0.84.0,<0.85.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "google-genai (>=1.66.0,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "fastmcp (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
